move files + connect LandingPage with FoamTreeSearch via routes
