+++ src/components/ElasticSearchPage/applyDisjunctiveFaceting.js
