+++ src/components/ElasticSearchPage/buildStateFacets.js
